Clean up mocks
