+++ src/xcode/ENA/ENA/Source/Services/HealthCertificateValidation/MockHealthCertificateValidationService.swift
@@ -1,34 +1,14 @@
-////
+//
 // 🦠 Corona-Warn-App
 //
 
 import Foundation
 import OpenCombine
-import HealthCertificateToolkit
-import SwiftyJSON
-import class CertLogic.ValidationResult
-import class CertLogic.Rule
-import class CertLogic.Description
+@testable import ENA
 
 struct MockHealthCertificateValidationService: HealthCertificateValidationProviding {
 
-	var onboardedCountriesResult: Result<[Country], HealthCertificateValidationOnboardedCountriesError> = .success(
-		[
-			Country(countryCode: "DE"),
-			Country(countryCode: "IT"),
-			Country(countryCode: "ES")
-		].compactMap { $0 }
-	)
-
-	var validationResult: Result<HealthCertificateValidationReport, HealthCertificateValidationError> = .success(.validationFailed([validationResult1, validationResult2]))
-
-	func onboardedCountries(
-		completion: @escaping (Result<[Country], HealthCertificateValidationOnboardedCountriesError>) -> Void
-	) {
-		DispatchQueue.main.asyncAfter(deadline: .now() + 0) {
-			completion(onboardedCountriesResult)
-		}
-	}
+	// MARK: - Protocol HealthCertificateValidationProviding
 	
 	func validate(
 		healthCertificate: HealthCertificate,
@@ -36,47 +16,11 @@ struct MockHealthCertificateValidationService: HealthCertificateValidationProvid
 		validationClock: Date,
 		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
 	) {
-		DispatchQueue.main.asyncAfter(deadline: .now() + 0) {
-			completion(validationResult)
-		}
+		completion(validationResult)
 	}
 
+	// MARK: - Internal
+
+	var validationResult: Result<HealthCertificateValidationReport, HealthCertificateValidationError> = .success(.validationPassed([]))
+
 }
-
-let validationResult1 = ValidationResult(
-	rule: Rule(
-		identifier: "TR-001",
-		type: "Acceptance",
-		version: "",
-		schemaVersion: "",
-		engine: "",
-		engineVersion: "",
-		certificateType: "",
-		description: [Description(lang: "de", desc: "Art des Tests muss Antigen-Test oder Nukleinsäureamplifikations-Test sein.")],
-		validFrom: "",
-		validTo: "",
-		affectedString: ["v.0.tg"],
-		logic: JSON(booleanLiteral: true),
-		countryCode: "IT"),
-	result: .fail,
-	validationErrors: []
-)
-
-let validationResult2 = ValidationResult(
-	rule: Rule(
-		identifier: "TR-002",
-		type: "Invalidation",
-		version: "",
-		schemaVersion: "",
-		engine: "",
-		engineVersion: "",
-		certificateType: "",
-		description: [Description(lang: "de", desc: "Ein Antigentest ist maximal 48h gültig.")],
-		validFrom: "",
-		validTo: "",
-		affectedString: ["t.0.ci"],
-		logic: JSON(booleanLiteral: true),
-		countryCode: "it"),
-	result: .open,
-	validationErrors: []
-)
